+++ ocrd/ocrd/cli/workspace.py
@@ -1,5 +1,5 @@
 import os
-from os.path import relpath, exists, join
+from os.path import relpath, exists, join, isabs
 from pathlib import Path
 import sys
 from glob import glob   # XXX pathlib.Path.glob does not support absolute globs
@@ -55,7 +55,7 @@ def workspace_cli(ctx, directory, mets_basename, backup):
 ''')
 @pass_workspace
 @click.option('-a', '--download', is_flag=True, help="Download all files")
-@click.option('-s', '--skip', help="Tests to skip", default=[], multiple=True, type=click.Choice(['imagefilename', 'dimension', 'mets_unique_identifier', 'mets_file_group_names', 'mets_files', 'pixel_density', 'page', 'url']))
+@click.option('-s', '--skip', help="Tests to skip", default=[], multiple=True, type=click.Choice(['imagefilename', 'dimension', 'mets_unique_identifier', 'mets_file_group_names', 'mets_files', 'pixel_density', 'page', 'page_xsd', 'mets_xsd', 'url']))
 @click.option('--page-textequiv-consistency', '--page-strictness', help="How strict to check PAGE multi-level textequiv consistency", type=click.Choice(['strict', 'lax', 'fix', 'off']), default='strict')
 @click.option('--page-coordinate-consistency', help="How fierce to check PAGE multi-level coordinate consistency", type=click.Choice(['poly', 'baseline', 'both', 'off']), default='poly')
 @click.argument('mets_url', nargs=-1)
@@ -154,15 +154,16 @@ def workspace_add_file(ctx, file_grp, file_id, mimetype, page_id, ignore, check_
 
     kwargs = {'fileGrp': file_grp, 'ID': file_id, 'mimetype': mimetype, 'pageId': page_id, 'force': force, 'ignore': ignore}
     log = getLogger('ocrd.cli.workspace.add')
+    log.debug("Adding '%s' (%s)", fname, kwargs)
     if not (fname.startswith('http://') or fname.startswith('https://')):
         if not fname.startswith(ctx.directory):
-            if exists(join(ctx.directory, fname)):
+            if not isabs(fname) and exists(join(ctx.directory, fname)):
                 fname = join(ctx.directory, fname)
             else:
                 log.debug("File '%s' is not in workspace, copying", fname)
                 try:
                     fname = ctx.resolver.download_to_directory(ctx.directory, fname, subdir=file_grp)
-                except FileNotFoundError as e:
+                except FileNotFoundError:
                     if check_file_exists:
                         log.error("File '%s' does not exist, halt execution!" % fname)
                         sys.exit(1)