+++ ocrd/ocrd/lib.bash
@@ -4,8 +4,56 @@
     exit 1
 
 # END-INCLUDE 
-# BEGIN-INCLUDE ./src/raise.bash 
+# BEGIN-INCLUDE ./src/logging.bash 
+## ### `ocrd__raise`
+## 
+## Raise an error and exit.
+ocrd__raise () {
+    echo >&2 "ERROR: $1"; exit 127
+}
 
+## ### `ocrd__log`
+## 
+## Delegate logging to `ocrd log`
+ocrd__log () {
+    local log_level="${ocrd__argv[log_level]:-}"
+    if [[ -n "$log_level" ]];then
+        ocrd -l "$log_level" log "$@"
+    else
+        ocrd log "$@"
+    fi
+}
+
+
+## ### `ocrd__minversion`
+## 
+## Ensure minimum version
+# ht https://stackoverflow.com/posts/4025065
+ocrd__minversion () {
+    set -x
+    local minversion="$1"
+    local version=$(ocrd --version|sed 's/ocrd, version //')
+    echo "$minversion < $version?"
+    if [[ $minversion == $version ]];then
+        return 0
+    fi
+    local IFS=.
+    version=($version)
+    minversion=($minversion)
+    # fill empty fields in version with zeros
+    for ((i=${#version[@]}; i<${#minversion[@]}; i++));do
+        version[i]=0
+    done
+    for ((i=0; i<${#version[@]}; i++));do
+        if [[ -z ${minversion[i]} ]];then
+            # fill empty fields in minversion with zeros
+            minversion[i]=0
+        fi
+        if ((10#${version[i]} < 10#${minversion[i]}));then
+            ocrd__raise "ocrd/core is too old (${version[*]} < ${minversion[*]}). Please update OCR-D/core"
+        fi
+    done
+}
 
 # END-INCLUDE 
 # BEGIN-INCLUDE ./src/dumpjson.bash 
@@ -58,6 +106,8 @@ ocrd__parse_argv () {
         ocrd__raise "Must set \$params (declare -A params)"
     fi
 
+    ocrd__argv[overwrite]=false
+
     while [[ "${1:-}" = -* ]];do
         case "$1" in
             -l|--log-level) ocrd__argv[log_level]=$2 ; shift ;;
@@ -69,6 +119,7 @@ ocrd__parse_argv () {
             -I|--input-file-grp) ocrd__argv[input_file_grp]=$2 ; shift ;;
             -w|--working-dir) ocrd__argv[working_dir]=$(realpath "$2") ; shift ;;
             -m|--mets) ocrd__argv[mets_file]=$(realpath "$2") ; shift ;;
+            --overwrite) ocrd__argv[overwrite]=true ;;
             -V|--version) ocrd ocrd-tool "$OCRD_TOOL_JSON" version; exit ;;
             *) ocrd__raise "Unknown option '$1'" ;;
         esac
@@ -87,14 +138,13 @@ ocrd__parse_argv () {
         ocrd__raise "log level '${ocrd__argv[log_level]}' is invalid"
     fi
 
-    # if [[ ! "${ocrd__argv[input_file_grp]:=OCR-D-IMG}" =~ OCR-D-(GT-)?(IMG|SEG|OCR|COR)(-[A-Z0-9\-]{3,})?(,OCR-D-(GT-)?(IMG|SEG|OCR|COR)(-[A-Z0-9\-]{3,})?)* ]];then
-    #     echo >&2 "WARNING: input fileGrp '${ocrd__argv[input_file_grp]}' does not conform to OCR-D spec"
-    # fi
+    if [[ -z "${ocrd__argv[input_file_grp]:=}" ]];then
+        ocrd__raise "Provide --input-file-grp/-I explicitly!"
+    fi
 
-    # if [[ ! "${ocrd__argv[output_file_grp]:=OCR-D-OCR}" =~ OCR-D-(GT-)?(IMG|SEG|OCR|COR)(-[A-Z0-9\-]{3,})?(,OCR-D-(GT-)?(IMG|SEG|OCR|COR)(-[A-Z0-9\-]{3,})?)* ]];then
-    #     echo >&2 "WARNING: output fileGrp '${ocrd__argv[output_file_grp]}' does not conform to OCR-D spec"
-    # fi
-
+    if [[ -z "${ocrd__argv[output_file_grp]:=}" ]];then
+        ocrd__raise "Provide --output-file-grp/-O explicitly!"
+    fi
 
     local params_parsed retval
     params_parsed="$(ocrd ocrd-tool "$OCRD_TOOL_JSON" tool $OCRD_TOOL_NAME parse-params -p "${ocrd__argv[parameter]:-{\}}")" || {
@@ -129,7 +179,7 @@ ocrd__wrap () {
         ocrd__raise "Cannot read \$OCRD_TOOL_JSON: '$OCRD_TOOL_JSON'"
     fi
 
-    if [[ -z "$OCRD_TOOL_NAME" ]];then
+   if [[ -z "$OCRD_TOOL_NAME" ]];then
         ocrd__raise "Must set \$OCRD_TOOL_NAME"
     elif ! ocrd ocrd-tool "$OCRD_TOOL_JSON" list-tools|grep -q "$OCRD_TOOL_NAME";then
         ocrd__raise "No such command \$OCRD_TOOL_NAME: $OCRD_TOOL_NAME"